# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OneLife Finance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/TransactionType"
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /emis:
    get:
      operationId: listEmis
      tags: [emis]
      summary: List EMIs across all loans
      parameters:
        - name: loanId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EmiStatus"
      responses:
        "200":
          description: List of EMIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Emi"
  /emis/{id}:
    patch:
      operationId: updateEmi
      tags: [emis]
      summary: Update an EMI (e.g. mark as paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmiUpdate"
      responses:
        "200":
          description: Updated EMI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emi"
        "404":
          description: Not found
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List budgets for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            description: "Month in YYYY-MM format"
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget line
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget line
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget line
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /credit-cards:
    get:
      operationId: listCreditCards
      tags: [credit_cards]
      summary: List credit cards
      responses:
        "200":
          description: List of credit cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditCard"
    post:
      operationId: createCreditCard
      tags: [credit_cards]
      summary: Create a credit card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditCardInput"
      responses:
        "201":
          description: Created credit card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCard"
  /credit-cards/{id}:
    patch:
      operationId: updateCreditCard
      tags: [credit_cards]
      summary: Update a credit card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditCardUpdate"
      responses:
        "200":
          description: Updated credit card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCard"
        "404":
          description: Not found
    delete:
      operationId: deleteCreditCard
      tags: [credit_cards]
      summary: Delete a credit card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    TransactionType:
      type: string
      enum: [income, expense]
    EmiStatus:
      type: string
      enum: [paid, pending, overdue, partial]
...
    Category:
      type: object
      required: [id, name, type, color, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type:
          $ref: "#/components/schemas/TransactionType"
        color: { type: string }
        icon: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CategoryInput:
      type: object
      required: [name, type, color]
      properties:
        name: { type: string, minLength: 1 }
        type:
          $ref: "#/components/schemas/TransactionType"
        color: { type: string, minLength: 1 }
        icon: { type: string }
...
    Emi:
      type: object
      required: [id, loanId, loanName, dueDate, paidDate, amount, status]
      properties:
        id: { type: integer }
        loanId: { type: integer }
        loanName: { type: string }
        dueDate: { type: string, format: date }
        paidDate: { type: ["string", "null"], format: date }
        amount: { type: number }
        status:
          $ref: "#/components/schemas/EmiStatus"
        penaltyAmount: { type: number }
        overdueDays: { type: integer }
    EmiUpdate:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/EmiStatus"
        paidDate: { type: ["string", "null"], format: date }
        penaltyAmount: { type: number }
        overdueDays: { type: integer }
...
    Budget:
      type: object
      required: [id, categoryId, categoryName, categoryColor, month, plannedAmount, actualAmount]
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        categoryName: { type: string }
        categoryColor: { type: string }
        month: { type: string }
        plannedAmount: { type: number }
        actualAmount: { type: number }
    BudgetInput:
      type: object
      required: [categoryId, month, plannedAmount]
      properties:
        categoryId: { type: integer }
        month: { type: string, minLength: 7 }
        plannedAmount: { type: number, minimum: 0 }
    BudgetUpdate:
      type: object
      properties:
        plannedAmount: { type: number, minimum: 0 }
...
    CreditCard:
      type: object
      required: [id, name, bankName, creditLimit, outstandingAmount, dueDate, minimumDue, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        bankName: { type: string }
        bankLogoUrl: { type: ["string", "null"] }
        creditLimit: { type: number }
        outstandingAmount: { type: number }
        dueDate: { type: string, format: date }
        minimumDue: { type: number }
        createdAt: { type: string, format: date-time }
    CreditCardInput:
      type: object
      required: [name, bankName, creditLimit, dueDate]
      properties:
        name: { type: string, minLength: 1 }
        bankName: { type: string, minLength: 1 }
        bankLogoUrl: { type: string }
        creditLimit: { type: number, minimum: 0 }
        outstandingAmount: { type: number, minimum: 0, default: 0 }
        dueDate: { type: string, format: date }
        minimumDue: { type: number, minimum: 0, default: 0 }
    CreditCardUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        bankName: { type: string, minLength: 1 }
        bankLogoUrl: { type: ["string", "null"] }
        creditLimit: { type: number, minimum: 0 }
        outstandingAmount: { type: number, minimum: 0 }
        dueDate: { type: string, format: date }
        minimumDue: { type: number, minimum: 0 }